Adding Sine and Cosine, Draft control
Sine wave and cosine wave shouldn't take the entire screen but I'll minimize it to fit the right side of the window.

The setters and getters are really important. I separated the views in two view classes: WaveView is what I'm supposed to do and WaveSettingsView is a draft version of what Brandon should have.

I recommend to keep the setters and getters BUT a menubar and control view pane should be made for the alpha build.

I will continue adding more graphs and elements like energy,etc. tomorrow.